Suppress logrus exit status error message (#309)
Addressing: https://github.com/Open-CMSIS-Pack/devtools/issues/1539

The purpose of this change is to clean up the `cbuild` error messages
that display the 'exit status X' after every command execution. To
suppress this, I’ve introduced a custom logger that wraps `logrus`,
modifying only the `Error` method. With this change, the default
severity level for this particular error will be set to `info`, meaning
they will only be visible when the `-v` (verbose) or `-d` (debug) flag
is used.